// src/constants/action-type.ts
export enum ActionType {
    MOVE_CELL = "MOVE_CELL",
    UPDATE_CELL = "UPDATE_CELL",
    DELETE_CELL = "DELETE_CELL",
    INSERT_CELL_AFTER = "INSERT_CELL_AFTER",
    BUNDLE_START = "BUNDLE_START",
    BUNDLE_COMPLETE = "BUNDLE_COMPLETE",
    FETCH_CELLS = 'FETCH_CELLS',
    FETCH_CELLS_COMPLETE = 'FETCH_CELLS_COMPLETE',
    FETCH_CELLS_ERROR = 'FETCH_CELLS_ERROR',
    SAVE_CELLS_ERROR = 'SAVE_CELLS_ERROR'
}